+++ common/src/main/java/com/st0x0ef/stellaris/common/blocks/entities/machines/OxygenDistributorBlockEntity.java
@@ -5,6 +5,8 @@ import com.fej1fun.potentials.fluid.UniversalFluidStorage;
 import com.fej1fun.potentials.providers.FluidProvider;
 import com.st0x0ef.stellaris.common.menus.OxygenDistributorMenu;
 import com.st0x0ef.stellaris.common.network.packets.SyncFluidPacketWithoutDirection;
+import com.st0x0ef.stellaris.common.oxygen.DimensionOxygenManager;
+import com.st0x0ef.stellaris.common.oxygen.GlobalOxygenManager;
 import com.st0x0ef.stellaris.common.registry.BlockEntityRegistry;
 import com.st0x0ef.stellaris.common.registry.FluidRegistry;
 import com.st0x0ef.stellaris.common.utils.capabilities.fluid.FluidUtil;
@@ -16,6 +18,7 @@ import net.minecraft.core.Direction;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.level.block.state.BlockState;
@@ -52,6 +55,13 @@ public class OxygenDistributorBlockEntity extends BaseEnergyContainerBlockEntity
     @Override
     public void tick() {
         FluidUtil.moveFluidFromItem(0, 0, items, oxygenTank, 5);
+
+        if (level instanceof ServerLevel serverLevel) {
+            DimensionOxygenManager oxygenManager = GlobalOxygenManager.getInstance().getOrCreateDimensionManager(serverLevel);
+            if (oxygenManager.getOxygenRoom(getBlockPos()) == null) {
+                oxygenManager.addOxygenRoom(getBlockPos());
+            }
+        }
     }
 
     public boolean useOxygenAndEnergy() {