+++ common/src/main/java/com/st0x0ef/stellaris/common/config/CommonConfig.java
@@ -2,8 +2,6 @@ package com.st0x0ef.stellaris.common.config;
 
 
 public class CommonConfig {
-    public float oxygenDamage = 2f;
-
     public int dieselGeneratorFuelTime = 20;
     public double orbitTeleportationYCoord = -10;
 
@@ -17,6 +15,9 @@ public class CommonConfig {
     @ConfigManager.InnerConfig
     public Gravity gravityConfig = new Gravity();
 
+    @ConfigManager.InnerConfig
+    public Oxygen oxygenConfig = new Oxygen();
+
     public static class Oil {
         public int chunkOilChance = 16;
 
@@ -27,6 +28,12 @@ public class CommonConfig {
     public static class Gravity {
         public boolean customEntityGravity = true;
         public boolean customItemGravity = true;
+    }
 
+    public static class Oxygen {
+        public boolean enableOxygenSystem = true;
+        public int oxygenCheckInterval = 20;
+        public int maxOxygenRoomSize = 32;
+        public float oxygenDamage = 2f;
     }
 }