+++ changelog.md
@@ -1,4 +1,8 @@
 # 1.4.9 changelog
 
+## Additions
+- New config options for the oxygen system
+
 ## Fix
 - Water separator crash
+- Fix some oxygen related issue